Implemented the first stages of the protocol.
The proposals are being exchanged. Now what's left is for
the buyer to accept the best proposal. Then, the seller will
decide whether to effectively supply or not based on some
random probability. If the seller doesn't comply, the 
buyer will remember that and inform the Monitor (TODO:
create a monitor) which will place that information in
the computational trust repository (TODO: create a computational
trust repository).